+++ src/TriviaRefactoringKata/TriviaRefactoringKata/Main.cpp
@@ -5,7 +5,7 @@ static bool notAWinner;
 
 int main(int argc, char *argv[])
 {
-	srand(time(NULL));
+	srand((int)time(NULL));
 
 	Game aGame;
